+++ resources/js/react/routes.js
@@ -1,0 +1,22 @@
+import React from "react";
+import { Switch, Route } from "react-router-dom";
+
+import Noticias from "./pages/admin/Noticias";
+import Membros from "./pages/admin/Membros";
+import Projetos from "./pages/admin/Projetos";
+import VisualizarProjeto from "./pages/admin/VisualizarProjeto";
+import NovoProjeto from "./pages/public/NovoProjeto";
+
+const Routes = () => {
+  const route = process.env.MIX_APP_ROUTE;
+  return (
+    <Switch>
+      <Route exact path={`${route}/administrativa/noticias`} component={Noticias} />
+      <Route exact path={`${route}/administrativa/membros`} component={Membros} />
+      <Route exact path={`${route}/administrativa/projetos`} component={Projetos} />
+      <Route exact path={`${route}/administrativa/projetos/:id`} component={VisualizarProjeto} />
+      <Route exact path={`${route}/projeto/novo`} component={NovoProjeto} />
+    </Switch>
+  );
+};
+export default Routes;
