+++ resources/js/react/styles/global.js
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from "styled-components";
+import "react-toastify/dist/ReactToastify.css";
+
+export default createGlobalStyle`
+
+#admin-area {
+    text-rendering: optimizelegibility;
+    -webkit-font-smoothing: antialiased !important;
+}
+
+`;
